fix(synapse-interface): two step approvals [SYN-62] (#3562)
* refactor: isolate submit approval tx

* feat: two-step approvals

* feat: fire toast when resetting allowance

* feat: add translation for resetting allowance

* fix: dismiss resetting allowance toast on error
