+++ packages/synapse-interface/messages/ar.json
@@ -246,6 +246,7 @@
       "Total Liquidity": "إجمالي السيولة",
       "Total Liquidity USD": "إجمالي السيولة بالدولار الأمريكي",
       "Requesting approval": "طلب الموافقة",
+      "Resetting allowance to zero first": "إعادة تعيين السماح إلى الصفر أولاً",
       "Starting your deposit": "بدء الإيداع…",
       "Starting your claim": "بدء المطالبة…"
     }
