+++ packages/synapse-interface/messages/en-US.json
@@ -246,6 +246,7 @@
       "Total Liquidity": "Total Liquidity",
       "Total Liquidity USD": "Total Liquidity USD",
       "Requesting approval": "Requesting approval",
+      "Resetting allowance to zero first": "Resetting allowance to zero first",
       "Starting your deposit": "Starting your deposit...",
       "Starting your claim": "Starting your claim..."
     }
