+++ packages/synapse-interface/messages/es.json
@@ -246,6 +246,7 @@
       "Total Liquidity": "Liquidez Total",
       "Total Liquidity USD": "Liquidez Total en USD",
       "Requesting approval": "Solicitando aprobación",
+      "Resetting allowance to zero first": "Restableciendo permiso a cero primero",
       "Starting your deposit": "Iniciando tu depósito...",
       "Starting your claim": "Iniciando tu reclamo..."
     }
