+++ packages/synapse-interface/messages/fr.json
@@ -246,6 +246,7 @@
       "Total Liquidity": "Liquidité totale",
       "Total Liquidity USD": "Liquidité totale USD",
       "Requesting approval": "Demande d'approbation",
+      "Resetting allowance to zero first": "Réinitialisation préalable de l'autorisation à zéro",
       "Starting your deposit": "Début du dépôt...",
       "Starting your claim": "Début de la réclamation..."
     }
