+++ packages/synapse-interface/messages/tr.json
@@ -246,6 +246,7 @@
       "Total Liquidity": "Toplam Likidite",
       "Total Liquidity USD": "Toplam Likidite USD",
       "Requesting approval": "Onay isteniyor",
+      "Resetting allowance to zero first": "İzin miktarı önce sıfırlanıyor",
       "Starting your deposit": "Yatırmanız başlatılıyor...",
       "Starting your claim": "Talebiniz başlatılıyor..."
     }
